+++ docs-src/src/templates/shortcodes/index.js
@@ -2,7 +2,7 @@ import popoverGridItem from "./popoverGridItem.js";
 import tileGridItem from "./tileGridItem.js";
 import scrollSlider from "./scrollSlider.js";
 import slider from "./slider.js";
-import card from "./card.js";
+import demoCard from "./demoCard.js";
 import flipcard from "./flipcard.js";
 import demoDataTable from "./demoDataTable.js";
 import demoCaptionFigure from "./demoCaptionFigure.js";
@@ -10,19 +10,21 @@ import demoHero from "./demoHero.js";
 import spokeSpinner from "./spokeSpinner.js";
 import fullscreenIntertitle from "./fullscreenIntertitle.js";
 import headline from "./headline.js";
+import demoCardGrid from "./demoCardGrid.js";
 
 export const shortcodes = {
   popoverGridItem,
   tileGridItem,
   scrollSlider,
-  card,
+  demoCard,
   slider,
   flipcard,
   demoCaptionFigure,
   demoDataTable,
   demoHero,
   spokeSpinner,
-  headline
+  headline,
+  demoCardGrid
 };
 
 export const pairedShortcodes = {